Fixed an issue with output paths not being computed correctly.
