+++ micall_docker.py
@@ -24,6 +24,7 @@ from micall.monitor.find_groups import find_groups
 from micall.monitor import error_metrics_parser, quality_metrics_parser
 from micall.g2p.pssm_lib import Pssm
 from micall.monitor.tile_metrics_parser import summarize_tiles
+from micall.utils.driver_utils import MiCallFormatter, safe_file_move, makedirs
 
 EXCLUDED_SEEDS = ['HLA-B-seed']  # Not ready yet.
 EXCLUDED_PROJECTS = ['HCV-NS5a',
@@ -294,44 +295,6 @@ if you want to write the outputs to "/host/output/path/":
         1234A_MIDI_forward.fastq 1234A_MIDI_reverse.fastq \\ 
         /data /results
 """
-
-
-class MiCallFormatter(ArgumentDefaultsHelpFormatter):
-    def _fill_text(self, text, width, indent):
-        import textwrap
-        lines = text.splitlines(keepends=True)
-        wrapped = ''
-        paragraph = ''
-        is_done = False
-        while not is_done:
-            if not lines:
-                is_done = True
-                is_blank = False
-                extra_indent = next_line = ''
-            else:
-                next_line = lines.pop(0)
-                indent_match = re.match(r'\s+', next_line)
-                extra_indent = indent_match.group(0) if indent_match else ''
-                is_blank = extra_indent == next_line
-                if is_blank:
-                    extra_indent = ''
-            if (is_done or is_blank or extra_indent) and paragraph:
-                paragraph = self._whitespace_matcher.sub(' ', paragraph).strip()
-                wrapped += textwrap.fill(paragraph,
-                                         width,
-                                         initial_indent=indent,
-                                         subsequent_indent=indent)
-                wrapped += os.linesep
-                paragraph = ''
-            if extra_indent:
-                wrapped += textwrap.fill(next_line,
-                                         width,
-                                         initial_indent=indent,
-                                         subsequent_indent=indent+extra_indent)
-                wrapped += os.linesep
-            elif not is_done and not is_blank:
-                paragraph += next_line
-        return wrapped
 
 
 def parse_args():
@@ -961,9 +924,12 @@ class MiCallArgs:
         if resolved_path is None:
             return None
         if not os.path.isabs(resolved_path):
+            results = self.original_args["results_folder"]
+            if not os.path.isabs(results):
+                results = os.path.join(self.original_args["run_folder"], results)
             io_prefix = (self.original_args["run_folder"]
                          if arg_name in self.PREFIX_WITH_RUN_FOLDER
-                         else self.original_args["results_folder"])
+                         else results)
             resolved_path = os.path.join(io_prefix, resolved_path)
         return resolved_path
 
@@ -1580,39 +1546,6 @@ def collate_samples(run_info):
         pass
     zip_folder(run_info.output_path, 'resistance_reports')
     zip_folder(run_info.output_path, 'coverage_maps')
-
-
-def safe_file_move(src, dst):
-    """
-    Helper that attempts to move a file from src to dst.
-
-    Because os.rename may fail on certain platforms, we fall back to
-    the boneheaded way of copying-and-deleting.
-    :param src:
-    :param dst:
-    :return:
-    """
-    try:
-        os.rename(src, dst)
-    except OSError as e:
-        if e.errno == errno.EXDEV:
-            logger.debug(
-                "Failed to rename %s to %s; copying and deleting the original.",
-                src,
-                dst,
-            )
-            shutil.copy(src, dst)
-            os.remove(src)
-        else:
-            raise
-
-
-def makedirs(path):
-    try:
-        os.makedirs(path)
-    except OSError as exc:
-        if exc.errno != errno.EEXIST or not os.path.isdir(path):
-            raise
 
 
 # noinspection PyTypeChecker,PyUnresolvedReferences
